+++ data-transfer/src/main/java/it/eng/datatransfer/service/AbstractDataTransferService.java
@@ -6,10 +6,7 @@ import it.eng.datatransfer.exceptions.TransferProcessInternalException;
 import it.eng.datatransfer.exceptions.TransferProcessInvalidFormatException;
 import it.eng.datatransfer.exceptions.TransferProcessInvalidStateException;
 import it.eng.datatransfer.exceptions.TransferProcessNotFoundException;
-import it.eng.datatransfer.model.TransferProcess;
-import it.eng.datatransfer.model.TransferRequestMessage;
-import it.eng.datatransfer.model.TransferStartMessage;
-import it.eng.datatransfer.model.TransferState;
+import it.eng.datatransfer.model.*;
 import it.eng.datatransfer.repository.TransferProcessRepository;
 import it.eng.datatransfer.repository.TransferRequestMessageRepository;
 import it.eng.datatransfer.serializer.TransferSerializer;
@@ -239,6 +236,125 @@ public abstract class AbstractDataTransferService implements TransferProcessStra
                         "consumerPid", transferProcessStarted.getConsumerPid(),
                         "providerPid", transferProcessStarted.getProviderPid()));
         return transferProcessStarted;
+    }
+
+    /**
+     * Finds transfer process, check if status is correct, publish event and update state to COMPLETED.
+     *
+     * @param transferCompletionMessage TransferCompletionMessage
+     * @param consumerPid               consumerPid in case of consumer callback usage
+     * @param providerPid               providerPid in case of provider usage
+     * @return TransferProcess with status COMPLETED
+     */
+    public TransferProcess completeDataTransfer(TransferCompletionMessage transferCompletionMessage, String consumerPid,
+                                                String providerPid) {
+        String consumerPidFinal = consumerPid == null ? transferCompletionMessage.getConsumerPid() : consumerPid;
+        String providerPidFinal = providerPid == null ? transferCompletionMessage.getProviderPid() : providerPid;
+        log.debug("Completing data transfer for consumerPid {} and providerPid {}", consumerPidFinal, providerPidFinal);
+
+        TransferProcess transferProcessStarted = findTransferProcess(consumerPidFinal, providerPidFinal);
+        stateTransitionCheck(transferProcessStarted, TransferState.COMPLETED);
+
+        TransferProcess transferProcessCompleted = TransferProcess.Builder.newInstance()
+                .id(transferProcessStarted.getId())
+                .agreementId(transferProcessStarted.getAgreementId())
+                .consumerPid(transferProcessStarted.getConsumerPid())
+                .providerPid(transferProcessStarted.getProviderPid())
+                .callbackAddress(transferProcessStarted.getCallbackAddress())
+                .dataAddress(transferProcessStarted.getDataAddress())
+                .isDownloaded(true)
+                .dataId(transferProcessStarted.getId())
+                .format(transferProcessStarted.getFormat())
+                .state(TransferState.COMPLETED)
+                .role(transferProcessStarted.getRole())
+                .datasetId(transferProcessStarted.getDatasetId())
+                .created(transferProcessStarted.getCreated())
+                .createdBy(transferProcessStarted.getCreatedBy())
+                .modified(transferProcessStarted.getModified())
+                .lastModifiedBy(transferProcessStarted.getLastModifiedBy())
+                .version(transferProcessStarted.getVersion())
+                .build();
+
+        saveTransferProcess(transferProcessCompleted);
+        publisher.publishEvent(TransferProcessChangeEvent.Builder.newInstance()
+                .oldTransferProcess(transferProcessStarted)
+                .newTransferProcess(transferProcessCompleted)
+                .build());
+        publisher.publishEvent(transferCompletionMessage);
+        publisher.publishEvent(AuditEventType.PROTOCOL_TRANSFER_COMPLETED,
+                "Transfer process completed",
+                Map.of("role", IConstants.ROLE_PROTOCOL,
+                        "transferProcess", transferProcessCompleted,
+                        "consumerPid", transferProcessCompleted.getConsumerPid(),
+                        "providerPid", transferProcessCompleted.getProviderPid()));
+        return transferProcessCompleted;
+    }
+
+    /**
+     * Transition data transfer to TERMINATED state.
+     *
+     * @param transferTerminationMessage message
+     * @param consumerPid                consumerPid in case of consumer callback usage
+     * @param providerPid                providerPid in case of provider usage
+     * @return TransferProcess with status TERMINATED
+     */
+    public TransferProcess terminateDataTransfer(TransferTerminationMessage transferTerminationMessage, String consumerPid,
+                                                 String providerPid) {
+        String consumerPidFinal = consumerPid == null ? transferTerminationMessage.getConsumerPid() : consumerPid;
+        String providerPidFinal = providerPid == null ? transferTerminationMessage.getProviderPid() : providerPid;
+        log.debug("Terminating data transfer for consumerPid {} and providerPid {}", consumerPidFinal, providerPidFinal);
+
+        // can be in any state except TERMINATED
+        TransferProcess transferProcess = findTransferProcess(consumerPidFinal, providerPidFinal);
+        stateTransitionCheck(transferProcess, TransferState.TERMINATED);
+
+        TransferProcess transferProcessTerminated = transferProcess.copyWithNewTransferState(TransferState.TERMINATED);
+        saveTransferProcess(transferProcessTerminated);
+        publisher.publishEvent(TransferProcessChangeEvent.Builder.newInstance()
+                .oldTransferProcess(transferProcess)
+                .newTransferProcess(transferProcessTerminated)
+                .build());
+        publisher.publishEvent(transferTerminationMessage);
+        publisher.publishEvent(AuditEventType.PROTOCOL_TRANSFER_TERMINATED,
+                "Transfer process completed",
+                Map.of("role", IConstants.ROLE_PROTOCOL,
+                        "transferProcess", transferProcessTerminated,
+                        "consumerPid", transferProcessTerminated.getConsumerPid(),
+                        "providerPid", transferProcessTerminated.getProviderPid()));
+        return transferProcessTerminated;
+    }
+
+    /**
+     * Transition data transfer to SUSPENDED state.
+     *
+     * @param transferSuspensionMessage message
+     * @param consumerPid               consumerPid in case of consumer callback usage
+     * @param providerPid               providerPid in case of provider usage
+     * @return TransferProcess with status SUSPENDED
+     */
+    public TransferProcess suspendDataTransfer(TransferSuspensionMessage transferSuspensionMessage, String consumerPid, String providerPid) {
+        String consumerPidFinal = consumerPid == null ? transferSuspensionMessage.getConsumerPid() : consumerPid;
+        String providerPidFinal = providerPid == null ? transferSuspensionMessage.getProviderPid() : providerPid;
+        TransferProcess transferProcess = findTransferProcess(consumerPidFinal, providerPidFinal);
+
+        log.info("Found TransferProcessState {} ", transferProcess.getState());
+        log.debug("Suspending data transfer for consumerPid {} and providerPid {}", consumerPidFinal, providerPidFinal);
+
+        stateTransitionCheck(transferProcess, TransferState.SUSPENDED);
+        log.info("Acting as consumer, suspend the transfer process");
+        TransferProcess transferProcessSuspended = transferProcess.copyWithNewTransferState(TransferState.SUSPENDED);
+        publisher.publishEvent(TransferProcessChangeEvent.Builder.newInstance()
+                .oldTransferProcess(transferProcess)
+                .newTransferProcess(transferProcessSuspended)
+                .build());
+        publisher.publishEvent(transferSuspensionMessage);
+        publisher.publishEvent(AuditEventType.PROTOCOL_TRANSFER_SUSPENDED,
+                "Transfer process completed",
+                Map.of("role", IConstants.ROLE_PROTOCOL,
+                        "transferProcess", transferProcessSuspended,
+                        "consumerPid", transferProcessSuspended.getConsumerPid(),
+                        "providerPid", transferProcessSuspended.getProviderPid()));
+        return saveTransferProcess(transferProcessSuspended);
     }
 
     /**